♻️ refactor client structure
