+++ findmyorder/handler/standard.py
@@ -15,7 +15,7 @@ from pyparsing import (
     pyparsing_common,
 )
 
-from .handler import ParserClient
+from ._client import ParserClient
 
 
 class StandardHandler(ParserClient):
